@@ -2,8 +2,19 @@
 
     <PropertyGroup>
         <TargetFramework>net8.0</TargetFramework>
-        <LangVersion>9</LangVersion>
+        <LangVersion>12</LangVersion>
         <Nullable>enable</Nullable>
+        <PackageId>SynesthesiaUtils</PackageId>
+        <Version>1.0</Version>
+        <Authors>SynesthesiaDev_</Authors>
+        <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+        <PackageReadmeFile>README.md</PackageReadmeFile>
+        <PackageLicenseFile>LICENSE</PackageLicenseFile>
+        <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
     </PropertyGroup>
+    <ItemGroup>
+        <None Include="README.md" Pack="true" PackagePath="\"/>
+        <None Include="LICENSE" Pack="true" PackagePath="\"/>
+    </ItemGroup>
 
 </Project>
